requirements.txt
# Verzija: 2.0 | Cloud deployment | Ažurirano: 2025-04-27

# UI
streamlit==1.45.0

# Parsiranje dokumenata
pypdf==5.4.0
python-docx==1.1.2

# Chunking
langchain-text-splitters==0.3.8

# Supabase klijent
supabase==2.15.1

# LLM — Groq API
groq==0.25.0

# HTTP pozivi (za HuggingFace Inference API)
httpx==0.28.1

# Konfiguracija
python-dotenv==1.1.0
